refactor: modularise embrace attribute class
