+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/logs/attachments/Attachment.kt
@@ -1,6 +1,5 @@
 package io.embrace.android.embracesdk.internal.logs.attachments
 
-import io.embrace.android.embracesdk.internal.arch.schema.EmbraceAttributeKey
 import io.embrace.android.embracesdk.internal.logs.attachments.AttachmentErrorCode.ATTACHMENT_TOO_LARGE
 import io.embrace.android.embracesdk.internal.logs.attachments.AttachmentErrorCode.OVER_MAX_ATTACHMENTS
 import io.embrace.android.embracesdk.internal.logs.attachments.AttachmentErrorCode.UNKNOWN
@@ -8,6 +7,7 @@ import io.embrace.android.embracesdk.internal.opentelemetry.embAttachmentErrorCo
 import io.embrace.android.embracesdk.internal.opentelemetry.embAttachmentId
 import io.embrace.android.embracesdk.internal.opentelemetry.embAttachmentSize
 import io.embrace.android.embracesdk.internal.opentelemetry.embAttachmentUrl
+import io.embrace.android.embracesdk.internal.otel.attrs.EmbraceAttributeKey
 import io.embrace.android.embracesdk.internal.utils.toNonNullMap
 import java.util.UUID
