+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/telemetry/EmbraceTelemetryService.kt
@@ -1,8 +1,8 @@
 package io.embrace.android.embracesdk.internal.telemetry
 
 import io.embrace.android.embracesdk.internal.SystemInfo
-import io.embrace.android.embracesdk.internal.arch.schema.toEmbraceAttributeName
 import io.embrace.android.embracesdk.internal.isEmulator
+import io.embrace.android.embracesdk.internal.otel.attrs.toEmbraceAttributeName
 import io.embrace.android.embracesdk.internal.spans.toEmbraceUsageAttributeName
 import java.util.concurrent.ConcurrentHashMap
