+++ embrace-android-features/build.gradle.kts
@@ -16,6 +16,7 @@ dependencies {
     compileOnly(project(":embrace-android-infra"))
     compileOnly(project(":embrace-android-payload"))
     compileOnly(project(":embrace-internal-api"))
+    compileOnly(project(":embrace-android-otel"))
     compileOnly(platform(libs.opentelemetry.bom))
     compileOnly(libs.opentelemetry.api)
     compileOnly(libs.opentelemetry.sdk)
@@ -28,6 +29,7 @@ dependencies {
     testImplementation(project(":embrace-android-infra"))
     testImplementation(project(":embrace-android-payload"))
     testImplementation(project(":embrace-internal-api"))
+    testImplementation(project(":embrace-android-otel"))
     testImplementation(platform(libs.opentelemetry.bom))
     testImplementation(libs.opentelemetry.api)
     testImplementation(libs.opentelemetry.sdk)
