+++ embrace-android-otel/src/main/kotlin/io/embrace/android/embracesdk/internal/otel/attrs/EmbraceAttributeExt.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.arch.schema
+package io.embrace.android.embracesdk.internal.otel.attrs
 
 /**
  * Prefix added to all attribute keys for all attributes with a specific meaning to the Embrace platform
@@ -18,7 +18,7 @@ private const val EMBRACE_SESSION_PROPERTY_NAME_PREFIX = "emb.properties."
 /**
  * Return the appropriate internal Embrace attribute name given the current string
  */
-internal fun String.toEmbraceAttributeName(isPrivate: Boolean = false): String {
+fun String.toEmbraceAttributeName(isPrivate: Boolean = false): String {
     val prefix = if (isPrivate) {
         EMBRACE_PRIVATE_ATTRIBUTE_NAME_PREFIX
     } else {
@@ -29,4 +29,6 @@ internal fun String.toEmbraceAttributeName(isPrivate: Boolean = false): String {
 
 fun String.toSessionPropertyAttributeName(): String = EMBRACE_SESSION_PROPERTY_NAME_PREFIX + this
 
-fun String.isSessionPropertyAttributeName(): Boolean = startsWith(EMBRACE_SESSION_PROPERTY_NAME_PREFIX)
+fun String.isSessionPropertyAttributeName(): Boolean = startsWith(
+    EMBRACE_SESSION_PROPERTY_NAME_PREFIX
+)
