+++ embrace-android-otel/src/main/kotlin/io/embrace/android/embracesdk/internal/otel/attrs/FixedAttribute.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.arch.schema
+package io.embrace.android.embracesdk.internal.otel.attrs
 
 import io.embrace.android.embracesdk.internal.payload.Attribute
