+++ embrace-test-fakes/build.gradle.kts
@@ -20,6 +20,7 @@ dependencies {
     compileOnly(project(":embrace-android-features"))
     compileOnly(project(":embrace-android-delivery"))
     compileOnly(project(":embrace-internal-api"))
+    implementation(project(":embrace-android-otel"))
 
     compileOnly(platform(libs.opentelemetry.bom))
     compileOnly(libs.opentelemetry.api)
